Handle login and out differently.
Instead of displaying the room object and using useEffect to connect to the room, we now connect once we receive the token back from the server and only display the Room component once the room is connected.

This way we need to handle the status of the application after the form is submitted but before the room is joined. This is done with a connecting state, which disables the form in the Lobby component while the app is connecting. Once it is connected, the Room component is displayed.

Now the VideoChat component handles logout, intentionally stopping all tracks and then disconnecting from the room.